ref: Extract useVirtualListDimentionChange from useVirtualizedInspector;
